refactor(project-chat): extract action-event helpers from HomePageClient.tsx
