+++ services/aris-web/app/HomePageClient.tsx
@@ -67,6 +67,15 @@ import {
   isUserEvent,
 } from './sessions/[sessionId]/chat-screen/helpers';
 import { MarkdownContent } from './sessions/[sessionId]/chat-screen/center-pane/renderers/MarkdownContent';
+import {
+  isProjectActionEvent,
+  isProjectRunStatusEvent,
+  eventCommand,
+  projectActionPreview,
+  projectActionMeta,
+} from '@/components/project-chat/helpers/projectChatEvents';
+import { renderCommandTokens, commandTokenClass } from '@/components/project-chat/helpers/commandTokens';
+import { GitActionMark, DockerActionMark } from '@/components/project-chat/helpers/actionMarks';
 
 type ProjectView = 'overview' | 'chats' | 'chat' | 'files' | 'context';
 type ComposerMode = 'agent' | 'plan' | 'terminal';
@@ -1833,99 +1842,6 @@ function agentAvatarClass(agent: SessionSummary['agent'] | SessionChat['agent'])
   if (agent === 'gemini') return 'msg__avatar--gemini';
   if (agent === 'codex') return 'msg__avatar--codex';
   return 'msg__avatar--sys';
-}
-
-function isProjectActionEvent(event: UiEvent): boolean {
-  if (readEventRole(event) === 'user') return false;
-  if (event.action?.command || event.action?.path || event.parsed?.commands?.length) return true;
-  return event.kind === 'command_execution'
-    || event.kind === 'docker_execution'
-    || event.kind === 'exec_execution'
-    || event.kind === 'file_list'
-    || event.kind === 'file_read'
-    || event.kind === 'file_write'
-    || event.kind === 'git_execution'
-    || event.kind === 'run_execution'
-    || event.kind === 'think';
-}
-
-function isProjectRunStatusEvent(event: UiEvent): boolean {
-  if (readEventRole(event) === 'user') return false;
-  const runStatus = readUiEventRunStatus(event);
-  if (!runStatus) return false;
-  const streamEvent = typeof event.meta?.streamEvent === 'string' ? event.meta.streamEvent.trim().toLowerCase() : '';
-  return streamEvent === 'run_status' || /^run status:/i.test(event.body || event.title);
-}
-
-function eventCommand(event: UiEvent): string {
-  return event.action?.command
-    || event.parsed?.commands?.[0]
-    || event.action?.path
-    || event.result?.preview
-    || event.body
-    || event.title;
-}
-
-function commandTokenClass(token: string, tokenIndex: number): string {
-  if (tokenIndex === 0) return 'pc-action-token--bin';
-  if (/^[A-Z_][A-Z0-9_]*=/.test(token)) return 'pc-action-token--env';
-  if (/^-{1,2}[\w-]+/.test(token)) return 'pc-action-token--flag';
-  if (/^https?:\/\//.test(token)) return 'pc-action-token--url';
-  if (/^(?:\/|~\/|\.{1,2}\/)/.test(token)) return 'pc-action-token--path';
-  if (/^(?:&&|\|\||[|;])$/.test(token)) return 'pc-action-token--op';
-  if (/^\d+(?:\.\d+)?(?::\d+)?$/.test(token) || /:\d+/.test(token)) return 'pc-action-token--number';
-  return 'pc-action-token--arg';
-}
-
-function renderCommandTokens(command: string) {
-  let tokenIndex = 0;
-  return (command.match(/\s+|[^\s]+/g) ?? []).map((token, index) => {
-    if (/^\s+$/.test(token)) {
-      return <span key={`space-${index}`}>{token}</span>;
-    }
-    const className = commandTokenClass(token, tokenIndex);
-    tokenIndex += 1;
-    return <span key={`${className}-${index}`} className={`pc-action-token ${className}`}>{token}</span>;
-  });
-}
-
-function GitActionMark({ size = 12 }: { size?: number }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" role="img" aria-label="Git">
-      <path
-        d="M22.1 10.9 13.1 1.9a1.55 1.55 0 0 0-2.2 0L8.98 3.82l2.38 2.38a1.92 1.92 0 0 1 2.43 2.45l2.28 2.28a1.92 1.92 0 1 1-1.15 1.15l-2.13-2.13v5.6a1.92 1.92 0 1 1-1.58-.02V9.82a1.92 1.92 0 0 1-.9-2.55L7.85 4.81 1.9 10.76a1.55 1.55 0 0 0 0 2.2l9.14 9.14a1.55 1.55 0 0 0 2.2 0l8.86-8.86a1.65 1.65 0 0 0 0-2.34Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-function DockerActionMark({ size = 12 }: { size?: number }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" role="img" aria-label="Docker">
-      <path
-        d="M4.9 11.8h2.2V9.7H4.9v2.1Zm2.7 0h2.2V9.7H7.6v2.1Zm2.7 0h2.2V9.7h-2.2v2.1Zm2.7 0h2.2V9.7H13v2.1Zm-5.4-2.6h2.2V7.1H7.6v2.1Zm2.7 0h2.2V7.1h-2.2v2.1Zm2.7 0h2.2V7.1H13v2.1Zm0-2.6h2.2V4.5H13v2.1Zm9.1 5.1c-.5-.3-1.5-.4-2.3-.2-.1-.8-.6-1.5-1.3-2l-.5-.3-.3.5c-.4.7-.5 1.7-.1 2.4.2.4.5.8.9 1-1.2.7-3.2.7-3.5.7H2.2l-.1.5c-.2 1.4.1 2.6.8 3.5.8 1.1 2.1 1.7 3.8 1.7 3.7 0 6.5-1.7 8-4.8 1 .1 3.1.1 4.5-1.3.6 0 1.9-.1 2.8-1l.4-.4-.3-.3Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-function projectActionMeta(kind: UiEvent['kind']) {
-  if (kind === 'file_read') return { Icon: FileSearch, label: 'Read', tone: 'read' };
-  if (kind === 'file_write') return { Icon: FilePenLine, label: 'Write', tone: 'write' };
-  if (kind === 'file_list') return { Icon: FolderTree, label: 'List', tone: 'list' };
-  if (kind === 'think') return { Icon: Brain, label: 'Thinking', tone: 'think' };
-  if (kind === 'git_execution') return { Icon: GitActionMark, label: 'Git', tone: 'git' };
-  if (kind === 'docker_execution') return { Icon: DockerActionMark, label: 'Docker', tone: 'docker' };
-  return { Icon: TerminalSquare, label: 'Run', tone: 'run' };
-}
-
-function projectActionPreview(event: UiEvent): string {
-  const command = eventCommand(event);
-  const preview = event.result?.preview || event.body || event.title || '';
-  if (!preview || preview === command) return '';
-  return preview.trim();
 }
 
 function projectRunStatusMeta(runStatus: string) {
